ecommerce_price_tracker/requirements.txt
aiohttp==3.9.3
asyncio==3.4.3
beautifulsoup4==4.12.3
SQLAlchemy==2.0.25
pymysql==1.1.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
APScheduler==3.10.4
python-dotenv==1.0.0
Pillow==10.2.0
decorator==5.1.1
requests==2.31.0
numpy==1.26.3
